security: isolate trusted MCP discovery from public server actions
